refactor(query): remove dead seam wrappers and converge on canonical modules
