--- sdk/src/query/command-resolution.ts
@@ -1,10 +1,0 @@
-export {
-  resolveQueryCommand,
-  resolveQueryTokens,
-  type QueryCommandRegistryLike,
-  type QueryCommandResolution,
-  type QueryMatchMode,
-  type QueryResolutionSource,
-  explainQueryCommandNoMatch,
-  type QueryCommandNoMatch,
-} from './query-command-resolution-strategy.js';
