--- sdk/src/query/normalize-query-command.ts
@@ -1,1 +1,0 @@
-export { normalizeQueryCommand } from './query-command-resolution-strategy.js';
