--- sdk/src/query/policy-convergence.ts
@@ -1,6 +1,0 @@
-export {
-  QUERY_POLICY_SNAPSHOT,
-  QUERY_MUTATION_COMMAND_LIST,
-  TRANSPORT_RAW_COMMANDS,
-  isQueryMutationCommand,
-} from './query-policy-snapshot.js';
