--- sdk/src/query/query-registry-capability.ts
@@ -1,4 +1,0 @@
-export {
-  supportsMutationCommand,
-  supportsRawOutputCommand,
-} from './query-policy-capability.js';
